WIP VTContext
Restructured backends to extend a top-level virtual object
	(more backends to update)

New VTProxy registration which just checkpoints status info
about that proxy, but also registers to the VTContext for special handling
of the actual data

  VeloC doesn't appear to have a way to checkpoint an object as
just a single node (e.g. each checkpoint of a proxy element is
a node's iteration's element data, not the iteration's element data),
so recovering to different node positions is currently not possible
  Working on that before I start the steps for handling recovery of
the VT data in the VTContext. I may just use update the file backend
to the new virtual hierarchy and use that for more configurability for now.

  Checkpoints of status info seem very large right now, some 66M for each node
with only one collection and one objgroup being checkpointed. I have not yet
investigated that

  Another todo, attaching listeners to collection element migration events,
to send the updated status info alongside. Not an immediate concern.